refactor(python): extract summary/HTML helpers into _summary
Move the display concern (summary_data, repr_text, repr_html + html/flow
helpers) out of _facade.py (1850 -> 1561 lines) into a new _summary module.
Behavior-preserving: facade methods delegate; public surface, signatures and
bulk routing guards all green; shell.py and test_api.py repointed to the new
home (private helpers; assertions unchanged).
